[ANDROSDK-1964] Move OptionFields to network package
